TEST: Remove system vars and use test containers to get props. Relocation of some code for clarity.
