+++ server/src/intTest/kotlin/org/tctalent/server/repository/db/integrationhelp/BaseDBIntegrationTest.kt
@@ -16,17 +16,12 @@
 
 package org.tctalent.server.repository.db.integrationhelp
 
-import io.github.oshai.kotlinlogging.KotlinLogging
 import org.junit.jupiter.api.BeforeAll
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
 import org.springframework.test.context.DynamicPropertyRegistry
 import org.springframework.test.context.DynamicPropertySource
-import org.testcontainers.containers.PostgreSQLContainer
 import org.testcontainers.junit.jupiter.Testcontainers
-import org.testcontainers.utility.MountableFile
-
-private val logger = KotlinLogging.logger {}
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
@@ -34,71 +29,20 @@ private val logger = KotlinLogging.logger {}
 abstract class BaseDBIntegrationTest {
 
   companion object {
-    private val createDb = arrayOf("psql", "-U", "postgres", "-c", "CREATE DATABASE tctalent;")
-    private val createUserCommand =
-      arrayOf("psql", "-U", "postgres", "-c", "CREATE USER tctalent WITH PASSWORD 'tctalent';")
-    private val superUserCommand =
-      arrayOf("psql", "-U", "postgres", "-c", "ALTER USER tctalent WITH SUPERUSER;")
-    private val dumpFilePath: String =
-      System.getenv("TC_DB_DUMP_FILE_PATH") ?: "FAIL: ENV VARIABLE not set"
-    private val containerMountPath: String =
-      System.getenv("TC_CONTAINER_MOUNT_PATH") ?: "FAIL: ENV VARIABLE not set"
-    private val psqlCommand =
-      arrayOf("psql", "-d", "tctalent", "-U", "tctalent", "-f", containerMountPath)
-
-    val db: PostgreSQLContainer<*> =
-      PostgreSQLContainer("postgres:14")
-        .also {
-          logger.info { "Creating the postgres container." }
-          logger.info { "Dump file path: $dumpFilePath" }
-          logger.info { "Container mount path: $containerMountPath" }
-        }
-        .apply {
-          withDatabaseName("tctalent")
-          withUsername("tctalent")
-          withPassword("tctalent")
-          withReuse(true)
-        }
+    private val dbContainer = DBContainer
 
     @BeforeAll
     @JvmStatic
     fun startDBContainer() {
-      db.apply { start() }
-      createDbObjects()
-      loadDb()
-      importDumpFileToDatabase()
-    }
-
-    private fun importDumpFileToDatabase() {
-      logger.info { "Importing the database dump." }
-      db.apply { execInContainer(*psqlCommand) }
-      logger.info { "Done importing the database dump." }
-    }
-
-    private fun createDbObjects() {
-      logger.info { "Creating database objects" }
-      db.apply {
-        execInContainer(*createDb)
-        execInContainer(*createUserCommand)
-        execInContainer(*superUserCommand)
-      }
-      logger.info { "Database container started. DB and user created." }
-    }
-
-    private fun loadDb() {
-      logger.info { "Copying dump file to the container." }
-      db.apply { copyFileToContainer(MountableFile.forHostPath(dumpFilePath), containerMountPath) }
-      logger.info { "Dump file copied to the database" }
+      dbContainer.startDBContainer()
     }
 
     @DynamicPropertySource
     @JvmStatic
     fun registerDBContainer(registry: DynamicPropertyRegistry) {
-      registry.add("spring.datasource.url", db::getJdbcUrl)
-      registry.add("spring.datasource.username", db::getUsername)
-      registry.add("spring.datasource.password", db::getPassword)
+      dbContainer.registerDBContainer(registry)
     }
   }
 
-  fun isContainerInitialized(): Boolean = db.isRunning
+  fun isContainerInitialized(): Boolean = dbContainer.db.isRunning()
 }
