+++ server/src/intTest/resources/application.yml
@@ -1,8 +1,4 @@
 spring:
-  datasource:
-    url: jdbc:tc:postgresql:latest:///tctalent
-    username: tctalent
-    password: tctalent
   jpa:
     properties.hibernate:
       show_sql: false
