+++ server/src/intTest/resources/testcontainers.properties
@@ -1,0 +1,18 @@
+#
+# Copyright (c) 2024 Talent Beyond Boundaries.
+#
+# This program is free software: you can redistribute it and/or modify it under
+# the terms of the GNU Affero General Public License as published by the Free
+# Software Foundation, either version 3 of the License, or any later version.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License
+# for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see https://www.gnu.org/licenses/.
+#
+
+testcontainers.dump.location=/tctalent/dump.sql
+testcontainers.container.mount=/docker-entrypoint-initdb.d/dump.sql
